+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/query-management/models/query-list/query-list-data/query-list-data.model.ts
@@ -1,28 +1,26 @@
-import { PartyMessage } from '../../party-messages/party-message.model';
-import { PartyMessagesGroup } from '../../party-messages/party-messages-group.model';
+import { CaseQueriesCollection, QueryMessage } from '../../case-queries-collection.model';
 import { QueryListItem } from '../query-list-item/query-list-item.model';
 
 export class QueryListData {
   public partyName: string;
   public roleOnCase: string;
-  public partyMessages: QueryListItem[];
+  public queries: QueryListItem[];
 
-  constructor(partyMessagesGroup: PartyMessagesGroup) {
-    this.partyName = partyMessagesGroup.partyName;
-    this.roleOnCase = partyMessagesGroup.roleOnCase;
-
+  constructor(caseQueriesCollection: CaseQueriesCollection) {
+    this.partyName = caseQueriesCollection.partyName;
+    this.roleOnCase = caseQueriesCollection.roleOnCase;
     // get the parent messages (messages without parentId) and add the children to them
-    const parentMessages = partyMessagesGroup.partyMessages.filter((message: PartyMessage) => !message.parentId);
-    this.partyMessages = parentMessages.map((message: PartyMessage) => this.buildQueryListItem(message, partyMessagesGroup.partyMessages));
+    const parentMessages = caseQueriesCollection.caseMessages.filter((message: QueryMessage) => !message.value.parentId);
+    this.queries = parentMessages.map((message: QueryMessage) => this.buildQueryListItem(message, caseQueriesCollection.caseMessages));
   }
 
-  private buildQueryListItem(message: PartyMessage, allMessages: PartyMessage[]): QueryListItem {
+  private buildQueryListItem(message: QueryMessage, allMessages: QueryMessage[]): QueryListItem {
     const queryListItem = new QueryListItem();
     Object.assign(queryListItem, {
-      ...message,
+      ...message.value,
       children: allMessages
-        .filter((childMessage: PartyMessage) => childMessage.parentId === message.id)
-        .map((childMessage: PartyMessage) => this.buildQueryListItem(childMessage, allMessages))
+        .filter((childMessage: QueryMessage) => childMessage.value.parentId === message.value.id)
+        .map((childMessage: QueryMessage) => this.buildQueryListItem(childMessage, allMessages))
     });
     return queryListItem;
   }